+++ extensions/wrapper/wrapper/src/main/java/de/sovity/edc/ext/wrapper/api/ui/UiResource.java
@@ -20,6 +20,8 @@ import de.sovity.edc.ext.wrapper.api.ui.model.AssetPage;
 import de.sovity.edc.ext.wrapper.api.ui.model.AssetCreateRequest;
 import de.sovity.edc.ext.wrapper.api.ui.model.ContractAgreementPage;
 import de.sovity.edc.ext.wrapper.api.ui.model.ContractAgreementTransferRequest;
+import de.sovity.edc.ext.wrapper.api.ui.model.ContractDefinitionRequest;
+import de.sovity.edc.ext.wrapper.api.ui.model.ContractDefinitionPage;
 import de.sovity.edc.ext.wrapper.api.ui.model.IdResponseDto;
 import de.sovity.edc.ext.wrapper.api.ui.model.PolicyDefinitionPage;
 import de.sovity.edc.ext.wrapper.api.ui.model.TransferHistoryPage;
@@ -27,6 +29,7 @@ import de.sovity.edc.ext.wrapper.api.ui.pages.asset.AssetApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.ContractAgreementPageApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.ContractAgreementTransferApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.policy.PolicyDefinitionApiService;
+import de.sovity.edc.ext.wrapper.api.ui.pages.contracts.ContractDefinitionApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.transferhistory.TransferHistoryPageApiService;
 import de.sovity.edc.ext.wrapper.api.ui.pages.transferhistory.TransferHistoryPageAssetFetcherService;
 import io.swagger.v3.oas.annotations.Operation;
@@ -53,6 +56,7 @@ public class UiResource {
     private final TransferHistoryPageAssetFetcherService transferHistoryPageAssetFetcherService;
     private final AssetApiService assetApiService;
     private final PolicyDefinitionApiService policyDefinitionApiService;
+    private final ContractDefinitionApiService contractDefinitionApiService;
 
     @GET
     @Path("pages/contract-agreement-page")
@@ -114,6 +118,30 @@ public class UiResource {
         return assetApiService.deleteAsset(assetId);
     }
 
+    @GET
+    @Path("pages/contract-definition-page")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Operation(description = "Collect all data for Contract Definition Page")
+    public ContractDefinitionPage contractDefinitionPage() {
+        return new ContractDefinitionPage(contractDefinitionApiService.getContractDefinitions());
+    }
+
+    @POST
+    @Path("pages/contract-definition-page/contract-definitions")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Consumes(MediaType.APPLICATION_JSON)
+    @Operation(description = "Create a new Contract Definition")
+    public IdResponseDto createContractDefinition(ContractDefinitionRequest contractDefinitionRequest) {
+        return contractDefinitionApiService.createContractDefinition(contractDefinitionRequest);
+    }
+
+    @DELETE
+    @Path("pages/contract-definition-page/contract-definitions/{contractDefinitionId}")
+    @Produces(MediaType.APPLICATION_JSON)
+    @Operation(description = "Delete a Contract Definition")
+    public IdResponseDto deleteContractDefinition(@PathParam("contractDefinitionId") String contractDefinitionId) {
+        return contractDefinitionApiService.deleteContractDefinition(contractDefinitionId);
+    }
 
     @GET
     @Path("pages/policy/page/policies")